truvari: fix dependencies, using python3.pkgs.callPackage
The problem was with having exact dependencies not vesion X or greater
(see https://github.com/spiralgenetics/truvari/issues/46). This just
pulls in the patch for lower-bound dependencies, which can be removed on
version 1.3.5 onwards.

I have also changed it to use not the toplevel `callPackage` but
`python3.pkgs.callPackage so that the dependencies are visible as
function arguments